Fixed drawing rooms not being properly left when changing drawings

See merge request polytechnique-montr-al/log3900/21-3/equipe-111/log3900-111!107

+++ mobile/app/src/main/java/com/example/colorimagemobile/ui/home/fragments/gallery/GalleryDrawingFragment.kt
@@ -61,13 +61,28 @@ class GalleryDrawingFragment : Fragment(R.layout.fragment_gallery_drawing) {
         DrawingSocketService.joinRoom(roomName!!)
     }
 
-    override fun onDestroy() {
-        super.onDestroy()
-        DrawingSocketService.disconnect()
+    private fun leaveDrawingRoom(){
+        if(roomName == null) return
         DrawingSocketService.leaveRoom(roomName!!)
+        DrawingSocketService.disconnect()
 
         roomName = null
         sharedPreferencesService.removeItem(Constants.STORAGE_KEY.DRAWING_ROOM_ID)
+    }
+
+    override fun onPause() {
+        super.onPause()
+        this.leaveDrawingRoom()
+    }
+
+    override fun onStop(){
+        super.onStop()
+        this.leaveDrawingRoom()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        this.leaveDrawingRoom()
     }
 
     // dynamically add tools on sidebar
